SQL Resource and code files
Two folders have been uploaded: Completed and Resources.  Folder named 'Completed' contains SQL code for the main assignment and python code (in jupyter notebook) for the Bonus portion of the assignment.  This folder also contains Images ERD and images for the bonus section of the assignment.  The folder named 'Resources' contain the CSV files for the assignment.